Spike to create rolling files
Will be used on android spike app for recording gps coordinates.
Done in browser because easier than android dev.
Allocate storage space because needed on chrome.
Create the root dir if needed on startup and then read files in it.
Get the latest version for each file (in the format filename-d.txt) so
the next file with that name will not erase it.